Manually fix Rails/HelperInstanceVariable issues
This is simply done by passing the instance variables as arguments.
One case where this isn't so simple is the 'active_tab' method,
which gets and sets a variable of the same name. While there are
almost certainly better ways to solve the problem this method exists
to solve, it's not in the scope of this change (ignored).

This also moves a surprise class in the TabsHelper into its own file
in the models/ directory, which already has simple model classes
that are not used to store data. Ideally we should separate data
models from presentation ones, but that's not in scope here.
