+++ app/helpers/tabs_helper.rb
@@ -1,4 +1,5 @@
 module TabsHelper
+  # rubocop:disable Rails/HelperInstanceVariable
   def active_tab
     return @active_tab if @active_tab
 
@@ -6,6 +7,7 @@ module TabsHelper
     tab_name = [visited_tab] & %w[metadata tagging history admin related_external_links unpublish]
     @active_tab = tab_name.blank? ? Edition::Tab["edit"] : Edition::Tab[tab_name.first]
   end
+  # rubocop:enable Rails/HelperInstanceVariable
 
   def tab_link(tab, edition_path)
     link_to tab.title,
@@ -18,43 +20,5 @@ module TabsHelper
 
   def tabs
     Edition::Tab.all
-  end
-
-  module Edition
-    class Tab
-      TABS = %w[edit tagging metadata history admin related_external_links unpublish].freeze
-      attr_accessor :name
-
-      def initialize(name = nil)
-        @name = name
-      end
-
-      def self.all
-        @all ||= TABS.map { |name| Tab.new(name) }
-      end
-
-      def self.[](name)
-        Tab.new(name)
-      end
-
-      def title
-        case name
-        when "history"
-          "History and notes"
-        when "related_external_links"
-          "Related external links"
-        else
-          name.capitalize
-        end
-      end
-
-      def path(edition_path)
-        name == "edit" ? edition_path : "#{edition_path}/#{name}"
-      end
-
-      def ==(other)
-        name == other.name
-      end
-    end
   end
 end
