+++ app/views/editions/show.html.erb
@@ -15,9 +15,9 @@
             <%= render 'link_check_reports/link_check_report', edition: @edition, report: @edition.latest_link_check_report %>
           </div>
 
-          <%= resource_form do |f| %>
+          <%= resource_form(@resource) do |f| %>
             <div class="well">
-              <%= render resource_fields, f: f %>
+              <%= render resource_fields(@resource), f: f %>
             </div>
 
             <%= edition_activities_fields(f, @resource) %>
